Refactor SpeecBotCard, ImageSlider, RecommendationssCarousel
