+++ DREAM Lab Frontend/src/App.css
@@ -3,85 +3,3 @@ body {
     background: linear-gradient(180deg, #042e55 0%, #7b7afd 100%);
     min-height: 100vh;
 }
-
-@media only screen and (min-height: 1025px) {
-    .inputRecomendaciones {
-        margin-top: 7rem;
-        margin-bottom: 4rem;
-    }
-}
-
-.inputRecomendaciones {
-    display: flex;
-    flex-wrap: nowrap;
-    justify-content: space-between; /* Change to space-between */
-    align-items: center; /* Center items vertically */
-    width: 60%;
-    margin-top: 12rem;
-    margin-bottom: 4rem;
-    margin-inline: auto;
-    padding: 1.3rem;
-    padding-left: 1.5rem;
-    padding-right: 1.5rem;
-}
-
-.text-field {
-    flex: 1; /* Change flex to 1 */
-    margin-left: 0.5rem;
-    margin-right: 0.5rem; /* Adjust margin */
-    border: none;
-    background-color: transparent;
-    color: white;
-    padding: 0.3rem;
-    outline: none;
-    font-size: 20px;
-}
-
-.microphone-button {
-    font-size: 2.5em;
-    padding-right: 60%;
-    padding-left: 30%;
-}
-
-.send-button {
-    width: 2rem;
-    margin-left: 0.5rem; /* Adjust margin */
-}
-
-@media only screen and (max-width: 700px) {
-    .inputRecomendaciones {
-        width: 100%;
-        flex-wrap: nowrap; /* Allow items to wrap on smaller screens */
-        justify-content: flex-start; /* Align items to start on smaller screens */
-    }
-
-    .microphone-button {
-        font-size: 2.2rem;
-        justify-content: center;
-    }
-
-    .text-field {
-        flex: 1;
-        width: 40%; /* Remove fixed width */
-    }
-    .send-button {
-        width: 2rem;
-    }
-}
-
-/* Animación micrófono activo */
-@keyframes pulse {
-    0% {
-        transform: scale(1);
-    }
-    50% {
-        transform: scale(0.8);
-    }
-    100% {
-        transform: scale(1);
-    }
-}
-
-.relative.active svg {
-    animation: pulse 0.5s infinite alternate;
-}
