+++ DREAM Lab Frontend/src/App.jsx
@@ -1,8 +1,8 @@
 import {
-  Route,
-  RouterProvider,
-  createBrowserRouter,
-  createRoutesFromElements,
+    Route,
+    RouterProvider,
+    createBrowserRouter,
+    createRoutesFromElements,
 } from "react-router-dom";
 
 import "./App.css";
@@ -23,7 +23,7 @@ import LoginPage from "./Login/LoginPage/LoginPage.jsx";
 import LandingPage from "./LandingPage/LandingPage.jsx";
 
 // Home
-import HomePage from "./Home/HomePage.jsx";
+import HomePage from "./pages/HomePage/HomePage.jsx";
 
 // Reservaciones
 import ReservacionSala from "./Reservaciones/ReservacionSala.jsx";
@@ -33,47 +33,54 @@ import SelectorEquipo from "./Reservaciones/SelectorEquipo.jsx";
 import SelectorSala from "./Reservaciones/SelectorSala.jsx";
 import LandingPageDev from "./LandingPage/LandingPageDev.jsx";
 import Profile from "./Profile/Profile.jsx";
-import Logros from "./Profile/Logros.jsx"
-import ReservacionesActivas from "./Profile/ReservacionesActivas.jsx"
+import Logros from "./Profile/Logros.jsx";
+import ReservacionesActivas from "./Profile/ReservacionesActivas.jsx";
 
 function secured(Component) {
-  return function WrappedComponent(props) {
-    return (
-      <ProtectedRoutes>
-        <Component {...props} />
-      </ProtectedRoutes>
-    );
-  };
+    return function WrappedComponent(props) {
+        return (
+            <ProtectedRoutes>
+                <Component {...props} />
+            </ProtectedRoutes>
+        );
+    };
 }
 
 const router = createBrowserRouter(
-  createRoutesFromElements(
-    <Route path="/" element={<Root />}>
-      <Route index element={<LandingPage />} />
-      <Route path="landingpage" element={<LandingPageDev />} /> {/* ruta provisional para desarrollo de la landing */}
-      <Route path="login" element={<LoginPage />} />
-      <Route path="home" element={secured(HomePage)()} />
-      <Route path="reservacion" element={secured(ReservacionSala)()} />
-      <Route
-        path="reservacion/confirmacion"
-        element={secured(Confirmacion)()}
-      />
-      <Route
-        path="reservacion/resumen"
-        element={secured(ResumenReservacion)()}
-      />
-      <Route path="reservacion/equipo" element={secured(SelectorEquipo)()} />
-      <Route path="reservacion/sala" element={secured(SelectorSala)()} />
-      <Route path="profile" element={secured(Profile)()} />
-      <Route path="*" element={<NotFound />} />
-      <Route path="/profile/logros" element={<Logros />} />
-      <Route path="/profile/reservaciones" element={<ReservacionesActivas />} />
-    </Route>
-  )
+    createRoutesFromElements(
+        <Route path="/" element={<Root />}>
+            <Route index element={<LandingPage />} />
+            <Route path="landingpage" element={<LandingPageDev />} />{" "}
+            {/* ruta provisional para desarrollo de la landing */}
+            <Route path="login" element={<LoginPage />} />
+            <Route path="home" element={secured(HomePage)()} />
+            <Route path="reservacion" element={secured(ReservacionSala)()} />
+            <Route
+                path="reservacion/confirmacion"
+                element={secured(Confirmacion)()}
+            />
+            <Route
+                path="reservacion/resumen"
+                element={secured(ResumenReservacion)()}
+            />
+            <Route
+                path="reservacion/equipo"
+                element={secured(SelectorEquipo)()}
+            />
+            <Route path="reservacion/sala" element={secured(SelectorSala)()} />
+            <Route path="profile" element={secured(Profile)()} />
+            <Route path="*" element={<NotFound />} />
+            <Route path="/profile/logros" element={<Logros />} />
+            <Route
+                path="/profile/reservaciones"
+                element={<ReservacionesActivas />}
+            />
+        </Route>
+    )
 );
 
 function App() {
-  return <RouterProvider router={router} />;
+    return <RouterProvider router={router} />;
 }
 
 export default App;
